% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excProbFun.R
\name{excProb}
\alias{excProb}
\title{Computes the exceedance probabilities on the raster of fine polygons}
\usage{
excProb(x, estimate, lemObjects, bw, threshold = 1, Nboot = 100,
  ncores = 2, tol = 1e-06, maxIter = 2000, filename = "")
}
\arguments{
\item{x}{Spatial polygons of case data}

\item{estimate}{Estimated intensity surface}

\item{lemObjects}{List of arrays for the smoothing matrix and 
raster stacks for the partition and smoothed offsets}

\item{bw}{bandwidth, specifying which smoothing matrix in \code{lemObjects} to use}

\item{threshold}{Vector of risk thresholds}

\item{Nboot}{Number of bootstraps}

\item{ncores}{Number of cores/threads for parallel processing}

\item{tol}{Tolerance for convergence}

\item{maxIter}{Maximum number of iterations}

\item{filename}{Passed to writeRaster}
}
\value{
The \code{excProb} function returns a raster stack of exceedance probabilities of specified risk thresholds.
}
\description{
The \code{excProb} function first bootstraps cases with the input risk thresholds, and then, 
 computes the exceedance probablities with the same bandwidth as the risk estimation on the cells of the fine raster.
}
\details{
After using the \code{excProb} function, the raster of exceedance probabilities is done on cells of the raster on the fine polygons.
}
\examples{
data(kentuckyCounty)

data(kentuckyTract)

\dontrun{
lemRaster = rasterPartition(polyCoarse = kentuckyCounty, polyFine = kentuckyTract,
                   cellsCoarse = 6, cellsFine = 100,
                   bw = c(10, 15, 20, 25) * 1000,
                   ncores = 4,
                   idFile = 'id.grd', offsetFile = 'offset.grd')

lemSmoothMat = smoothingMatrix(rasterObjects = lemRaster,
                   ncores = 4)

lemCv = lemXv(x = kentuckyCounty,
                   lemObjects = lemSmoothMat,
                   ncores = 4)
bestBw = lemCv$bw[which.min(lemCv$cv)]

lemRisk = riskEst(x = kentuckyCounty,
                   lemObjects = lemSmoothMat,
                   bw = bestBw)
                   
lemExcProb = excProb(x = kentuckyCounty, 
                   lemObjects = lemSmoothMat, 
					estimate=lemRisk,
                   threshold = c(1, 1.1, 1.25), 
                   Nboot = 100, 
                   ncores = 4) 
                   
plot(lemExcProb)
}

}

