% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kentuckyData.R
\docType{data}
\name{kentuckyCounty}
\alias{kentuckyCounty}
\title{Lung cancer cases in Kentucky}
\format{A SpatialPolygonsDataFrame of Kentucky county boundaries, populations and cancer cases.}
\usage{
data(kentuckyCounty)
}
\description{
The dataset contains lung cancer cases for the counties of Kentucky.
}
\details{
The lung cancer case data are obtained from www.cancer-rates.info and were simulated from a log Gaussian Cox process for deliberately unspecified years.
}
\examples{
data(kentuckyCounty)
spplot(kentuckyCounty, 'Cases')

}
\keyword{datasets}

