% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kentuckyData.R
\docType{data}
\name{kentuckyTract}
\alias{kentuckyTract}
\title{Population data in Kentucky}
\format{A SpatialPolygonsDataFrame of Kentucky census tract boundaries and populations.}
\usage{
data(kentuckyTract)
}
\description{
The dataset contains population data for the census tracts of Kentucky.
}
\details{
The lung cancer case data are obtained from www.cancer-rates.info and were simulated from a log Gaussian Cox process for deliberately unspecified years.
}
\examples{
data(kentuckyTract)
spplot(kentuckyTract, 'logExpected_surfaceArea')

}
\keyword{datasets}

