% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kentuckyData.R
\docType{data}
\name{kMap}
\alias{kMap}
\title{Map of Kentucky}
\format{A SpatialPolygonsDataFrame of Kentucky census tract boundaries and populations.}
\usage{
data('kMap')
}
\description{
Stamen toner map of Kentucky
}
\details{
The lung cancer case data are obtained from www.cancer-rates.info and were simulated from a log Gaussian Cox process for deliberately unspecified years.
}
\examples{
data('kentuckyTract')
data('kMap')
mapmisc::map.new(kentuckyTract)
plot(kentuckyTract, col='yellow', border='#00FF0020', add=TRUE)
plot(kMap, add=TRUE)
mapmisc::scaleBar(kentuckyTract, 'topleft', bg='white')
}
\keyword{datasets}

