% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemEstFun.R
\name{lemEst}
\alias{lemEst}
\title{Computes the relative risk estimation on the raster of fine polygons}
\usage{
lemEst(x, lemObjects, bw, ncores = 2, tol = 1e-06, maxIter = 2000,
  verbose = FALSE)
}
\arguments{
\item{x}{Spatial polygons of case data}

\item{lemObjects}{List of arrays for the smoothing matrix and
raster stacks for the partition and smoothed offsets}

\item{bw}{Numeric value of bandwidth}

\item{ncores}{Number of cores/threads for parallel processing}
}
\value{
The \code{lemEst} function returns a list containing the raster of risk estimations,
 array of smoothing matrix at the final iteration, and
 input rasters of partitions and smoothed offsets.
}
\description{
The \code{lemEst} function first creates the smoothing matrix at the final iteration with the input bandwidth, and then,
 computes the estimates of the relative risk on the cells of the fine raster.
}
\details{
After using the \code{lemEst} function, the raster of risk estimations is done on cells of the raster on the fine polygons.
}
\examples{
data(kentuckyCounty)

data(kentuckyTract)

\dontrun{
lemRaster = rasterPartition(polyCoarse = kentuckyCounty, polyFine = kentuckyTract,
                   cellsCoarse = 40, cellsFine = 400,
                   bw = c(10, 15, 20, 25) * 1000,
                   ncores = 4,
                   idFile = 'id.grd', offsetFile = 'offset.grd')

lemSmoothMat = smoothingMatrix(rasterObjects = lemRaster,
                   ncores = 4)

lemCv = lemXv(x = kentuckyCounty,
                   lemObjects = lemSmoothMat,
                   ncores = 4)

lemRisk = lemEst(x = kentuckyCounty,
                   lemObjects = lemSmoothMat,
                   bw = 15000,
                   ncores = 4)

plot(lemRisk$risk)
}

}

