% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalSmooth.R
\name{lemFinal}
\alias{lemFinal}
\title{Computes final smoothed LEM estimate}
\usage{
lemFinal(x, counts = colnames(x$xv)[-1], bw = x$xv[apply(x$xv[, counts], 2,
  which.min), "bw"], ncores = 1, filename = paste(tempfile(), ".grd", sep =
  ""), verbose = FALSE)
}
\arguments{
\item{x}{result from lemXV}

\item{counts}{observed counts, defaults to all counts for which cross validation was done}

\item{bw}{smoothing bandwidth, defaults to the optimal bandwidths for each set of counts}

\item{ncores}{Number of cores/threads for parallel processing}

\item{filename}{name of final raster file}

\item{verbose}{if TRUE, print progress information}
}
\value{
A raster brick
}
\description{
Takes a partition-level LEM estiamte and smoothis it to a continuous risk estimate
}
\details{
The optimal bandwidth for each layer is
}
