% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemXvFun.R
\name{lemXv}
\alias{lemXv}
\title{Computes the likelihood cross-validation}
\usage{
lemXv(cases, population, cellsCoarse, cellsFine, bw, xv = 4, lemObjects,
  ncores = 1, tol = 1e-06, maxIter = 2000, fact = 1,
  randomSeed = NULL, verbose = FALSE, path = getwd())
}
\arguments{
\item{cases}{spatial object with case counts}

\item{population}{spatial object with population data}

\item{cellsCoarse}{number of cells for parallel computations of the smoothing matrix}

\item{cellsFine}{number of cells for numerical integration}

\item{bw}{vector of bandwidths}

\item{xv}{Number of cross-validation datasets}

\item{lemObjects}{List of arrays for the smoothing matrix, and raster stacks for the partition and smoothed offsets}

\item{ncores}{Number of cores/threads for parallel processing}

\item{tol}{Tolerance for convergence}

\item{maxIter}{Maximum number of iterations for convergence}

\item{fact}{aggregation factor for offsets prior to smoothing}

\item{randomSeed}{Seed for random number generator}

\item{verbose}{Verbose output}

\item{path}{folder for storing rasters}
}
\value{
The \code{lemXv} function returns a data frame of specified bandwidths and their cross-validation scores.
}
\description{
The \code{lemXv} function computes the likelihood cross-validation scores for the observed data with the input bandwidths used for the smoothing matrix. The cross-valiation test and training datasets of the observed cases are generated by k-fold sampling without replacement.
}
\details{
After using the \code{lemXv} function, the last entry of the data frame object is the theoretical cross-validation score when bandwidth is set to infinity.
}
