% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterPartFun.R
\name{rasterPartition}
\alias{rasterPartition}
\title{Generates the partitions by overlaying rasters of the coarse and fine polygons}
\usage{
rasterPartition(polyCoarse, polyFine, cellsCoarse, cellsFine, bw,
  focalSize = NULL, fact = 1, xv = NULL, ncores = 1, path = getwd(),
  idFile = paste(tempfile("lemId", path), ".grd", sep = ""),
  offsetFile = paste(tempfile("lemOffset", path), ".grd", sep = ""),
  verbose = FALSE)
}
\arguments{
\item{polyCoarse}{Spatial polygons of case data}

\item{polyFine}{Spatial polygons of population data}

\item{cellsCoarse}{Horizontal/vertical resolution of raster applied to coarse polygons}

\item{cellsFine}{Horizontal/vertical resolution of raster applied to fine polygons}

\item{bw}{Vector of bandwidths}

\item{focalSize}{Distance to truncate Gaussian kernel, default is 2.5 times largest bandwidth}

\item{fact}{aggregation factor for offsets prior to smoothing}

\item{xv}{Number of cross-validation sets, or a matrix where rows are coarse polygons and columns are xv sets}

\item{ncores}{Number of cores/threads for parallel processing}

\item{path}{folder to store raster data}

\item{idFile}{Filename (must have .grd extension) of the raster of partitions}

\item{offsetFile}{Filename (must have .grd extension) of the rasters of smoothed offsets}

\item{verbose}{Verbose output}
}
\value{
The \code{rasterPartition} function returns a list containing the raster of the coarse polygons, raster stacks of the partitions and offsets, focal weight matrix of the Gaussian kernel, and the input coarse polygons.
}
\description{
The \code{rasterPartition} function first rasterizes the coarse and fine spatial polygons based on their respective input resolutions, and then, overlays these rasters to generate the raster of partitions of the local-EM algorithm. It also applies the kernel smoothing function with input bandwidths to the expected counts of the fine polygons to obtain the smoothed offsets (i.e., smoothed expected counts / cell area) of the partitions.
}
\details{
After using the \code{rasterPartition} function, the fine raster is a raster stack containing the IDs for the partitions created by overlaying the coarse and fine rasters. The offset raster is a raster stack containing the offsets of the partitions smoothed with the specified bandwidths. These values represent the denominator of the kernel smoothing matrix.
}
\examples{
\dontrun{ 
data(kentuckyCounty)
data(kentuckyTract)


lemRaster = rasterPartition(polyCoarse = kentuckyCounty, 
                           polyFine = kentuckyTract, 
                           cellsCoarse = 6, 
                           cellsFine = 100, 
                           bw = c(10, 15) * 1000, 
                           ncores = 2, 
                           path= tempdir(), 
                           verbose = TRUE)
}

}
