% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemEstFun.R
\name{riskEst}
\alias{riskEst}
\title{Computes the relative risk estimation}
\usage{
riskEst(x, lemObjects, bw, tol = 1e-06, maxIter = 2000)
}
\arguments{
\item{x}{Spatial polygons of case data}

\item{lemObjects}{List of arrays for the smoothing matrix, and raster stacks for the partition and smoothed offsets}

\item{bw}{Bandwidth specifying which smoothing matrix in \code{lemObjects} to use}

\item{tol}{Tolerance for convergence}

\item{maxIter}{Maximum number of iterations for convergence}
}
\value{
The \code{riskEst} function returns a raster of risk estimations for the input bandwidth.
}
\description{
The \code{riskEst} function computes the estimations of the relative risk with fine raster resolution.
}
\details{
After using the \code{riskEst} function, the risk estimations are computed on a fine resolution based on the rasterization of the spatial polygons of population data.
}
\examples{
\dontrun{ 
data('kentuckyCounty')
data('kentuckyTract')

lemRaster = rasterPartition(polyCoarse = kentuckyCounty, 
							polyFine = kentuckyTract, 
 	                        cellsCoarse = 6, 
                             cellsFine = 100, 
                             bw = c(10, 15) * 1000, 
                             ncores = 2, 
                             path = tempdir(), 
                             verbose = TRUE)


lemSmoothMat = smoothingMatrix(rasterObjects = lemRaster, 
                                ncores = 2, 
                                verbose = TRUE)

lemRisk = riskEst(x = kentuckyCounty, 
                   lemObjects = lemSmoothMat, 
                   bw = 15 * 1000) 

}

}
