% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excProbFun.R
\name{excProb}
\alias{excProb}
\title{Computes the exceedance probabilities on the raster of fine polygons}
\usage{
excProb(x, lemObjects, threshold = 1, Nboot = 100, ncores = 2,
  tol = 1e-06, maxIter = 2000, verbose = FALSE)
}
\arguments{
\item{x}{Spatial polygons of case data}

\item{lemObjects}{List of arrays for the smoothing matrix and 
raster stacks for the partition and smoothed offsets}

\item{threshold}{Vector of risk thresholds}

\item{Nboot}{Number of bootstraps}

\item{ncores}{Number of cores/threads for parallel processing}
}
\value{
The \code{excProb} function returns a raster stack of the risk estimation and 
 exceedance probabilities of specified risk thresholds.
}
\description{
The \code{excProb} function first bootstraps cases with the input risk thresholds, and then, 
 computes the exceedance probablities with the same bandwidth as the risk estimation on the cells of the fine raster.
}
\details{
After using the \code{excProb} function, the raster of exceedance probabilities is done on cells of the raster on the fine polygons.
}
\examples{
data(kentuckyCounty)

data(kentuckyTract)

\dontrun{
lemRaster = rasterPartition(polyCoarse = kentuckyCounty, polyFine = kentuckyTract, 
                   cellsCoarse = 40, cellsFine = 400, 
                   bw = c(10, 15, 20, 25) * 1000, 
                   ncores = 4, 
                   idFile = 'id.grd', offsetFile = 'offset.grd')
                   
lemSmoothMat = smoothingMatrix(rasterObjects = lemRaster, 
                   ncores = 4) 
                   
lemCv = lemXv(x = kentuckyCounty, 
                   lemObjects = lemSmoothMat, 
                   ncores = 4) 
                   
lemRisk = lemEst(x = kentuckyCounty, 
                   lemObjects = lemSmoothMat, 
                   bw = 15000, 
                   ncores = 4) 
                   
lemExcProb = excProb(x = kentuckyCounty, 
                   lemObjects = lemRisk, 
                   threshold = c(1, 1.1, 1.25), 
                   Nboot = 100, 
                   ncores = 4) 
                   
plot(lemExcProb)
}

}

