% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kentuckyData.R
\docType{data}
\name{kentuckyCounty}
\alias{kentuckyCounty}
\title{Observed cancer case data in Kentucky}
\format{A SpatialPolygonsDataFrame of Kentucky county boundaries and observed cancer cases.}
\usage{
data(kentuckyCounty)
}
\description{
The dataset contains simulated lung cancer cases for the counties of Kentucky.
}
\details{
The observed lung cancer case data are simulated from a log Gaussian Cox process with cancer rates obtained from www.cancer-rates.info for deliberately unspecified years.
}
\examples{
data(kentuckyCounty)

countyBrks = classInt::classIntervals(kentuckyCounty$count, 
		n = 7, style = 'quantile'
)
countyCol = classInt::findColours(countyBrks, 
		pal = rev(RColorBrewer::brewer.pal(7, 'RdYlBu'))
)

mapmisc::map.new(kentuckyCounty)
plot(kentuckyCounty, 
		col = countyCol, 
		axes = FALSE, border = FALSE
)

data(kMap)
plot(kMap, add = TRUE)

legend('topleft', 
		legend = rev(names(attr(countyCol, 'table'))), 
		fill = rev(attr(countyCol, 'palette')), 
		bg = 'white'
)
}
\keyword{datasets}
