% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kentuckyData.R
\docType{data}
\name{kMap}
\alias{kMap}
\title{Map of Kentucky}
\format{A RayerLayer of Kentucky and surrounding areas.}
\usage{
data('kMap')
}
\description{
Stamen toner map of Kentucky
}
\details{
The Stamen toner map is a detailed black and white background map to use with colorful overlays.
}
\examples{
data(kentuckyCounty)

countyBrks = classInt::classIntervals(kentuckyCounty$count, 
		n = 7, style = 'quantile'
)
countyCol = classInt::findColours(countyBrks, 
		pal = rev(brewer.pal(7, 'RdYlBu'))
)

mapmisc::map.new(kentuckyCounty)
plot(kentuckyCounty, 
		col = countyCol, 
		axes = FALSE, border = FALSE
)
plot(kMap, add = TRUE)

legend('topleft', 
		legend = rev(names(attr(countyCol, 'table'))), 
		fill = rev(attr(countyCol, 'palette')), 
		bg = 'white'
)

data(kentuckyTract)
kentuckyTract$expected[is.na(kentuckyTract$expected)] = 0

ctBrks = classInt::classIntervals(kentuckyTract$expected, 
	n = 7, style = 'fixed', fixedBreaks = c(0, 1, 1.5, 2, 3, 4, 6, 9.2)
)
ctCol = classInt::findColours(ctBrks, 
	pal = rev(brewer.pal(7, 'RdYlBu'))
)

mapmisc::map.new(kentuckyTract)
plot(kentuckyTract, 
	col = ctCol, 
	axes = FALSE, border = FALSE
)
plot(kMap, add = TRUE)

legend('topleft', 
		legend = rev(names(attr(ctCol, 'table'))), 
	fill = rev(attr(ctCol, 'palette')), 
	bg = 'white'
)
}
\keyword{datasets}

