% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalSmooth.R
\name{lemFinal}
\alias{lemFinal}
\title{Computes final smoothed LEM estimate}
\usage{
lemFinal(x, bw = NULL, ncores = 1, filename = tempfile(),
  verbose = FALSE)
}
\arguments{
\item{x}{result from lemXV}

\item{ncores}{Number of cores/threads for parallel processing}

\item{filename}{name of final raster file}
}
\value{
A raster brick
}
\description{
Takes a partition-level LEM estiamte and smoothis it to a continuous risk estimate
}
\details{
The optimal bandwidth for each layer is
}
