% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemXvFun.R
\name{lemXv}
\alias{lemXv}
\title{Computes the likelihood cross-validation}
\usage{
lemXv(x, lemObjects, Nxv = 5, ncores = 4, tol = 1e-06, maxIter = 2000,
  randomSeed = NULL, verbose = FALSE)
}
\arguments{
\item{\code{x}}{Spatial polygons of case data}

\item{\code{lemObjects}}{List of arrays for the smoothing matrix and 
raster stacks for the partition and smoothed offsets}

\item{\code{Nxv}}{Number of cross-validation datasets}

\item{\code{ncores}}{Number of cores/threads for parallel processing}
}
\value{
The \code{lemXv} function returns a data frame of specified bandwidths and cross-validation scores.
}
\description{
The \code{lemXv} function computes the likelihood cross-validation scores for the observed data with the input bandwidths used for the smoothing matrix. 
 The cross-valiation test and training datasets of the observed cases are generated by k-fold sampling without replacement.
}
\details{
After using the \code{lemXv} function, the last entry of the data frame object is the theoretical cross-validation score when bandwidth is set to infinity.
}
\examples{
data(kentuckyCounty)

data(kentuckyTract)

\dontrun{
lemRaster = rasterPartition(polyCoarse = kentuckyCounty, polyFine = kentuckyTract, 
                   cellsCoarse = 40, cellsFine = 400, 
                   bw = c(10, 15, 20, 25) * 1000, 
                   ncores = 4, 
                   idFile = 'id.grd', offsetFile = 'offset.grd')
                   
lemSmoothMat = smoothingMatrix(rasterObjects = lemRaster, 
                   ncores = 4) 
                   
lemCv = lemXv(x = kentuckyCounty, 
                   lemObjects = lemSmoothMat, 
                   ncores = 4) 
}

}

