% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemXvFun.R
\name{lemXv}
\alias{lemXv}
\title{Computes the likelihood cross-validation of the raster partitions}
\usage{
lemXv(cases, population, cellsCoarse, cellsFine, bw, xv = 4, lemObjects,
  ncores = 1, iterations = list(tol = 1e-05, maxIter = 1000, gpu = FALSE),
  randomSeed = NULL, path = getwd(), verbose = FALSE)
}
\arguments{
\item{cases}{Spatial polygons with case data}

\item{population}{Spatial polygons with population data}

\item{cellsCoarse}{Minimum resolution for rasterization of case data for numerical accuracy of smoothing matrix}

\item{cellsFine}{Minimum resolution for rasterization of population data for numerical integration of smoothing matrix}

\item{bw}{Vector of bandwidths}

\item{xv}{Number of cross-validation datasets}

\item{lemObjects}{List of arrays for the smoothing matrix, and raster stacks for the partition and smoothed offsets}

\item{ncores}{Number of cores/threads for parallel processing}

\item{iterations}{List of convergence tolerance, number of iterations, and use of gpuR package for running local-EM recursions}

\item{randomSeed}{Seed for random number generator}

\item{path}{Folder for storing rasters}

\item{verbose}{Verbose output}
}
\value{
The \code{lemXv} function returns a data frame of specified bandwidths and their cross-validation scores, and a raster of the risk estimation of the bandwidth with the lowest cross-validiation score.
}
\description{
The \code{lemXv} function computes the likelihood cross-validation scores for the observed data with the input bandwidths used for the smoothing matrix. The cross-valiation test and training datasets of the observed cases are generated by k-fold sampling without replacement.
}
\details{
After using the \code{lemXv} function, a raster stack containing the IDs for the partitions is created by overlaying the spatial polygons of the case and population data. The smoothed offsets and smoothing matrix are computed for the specified bandwidths of each cross-validation set.
}
