% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterPartFun.R
\name{rasterPartition}
\alias{rasterPartition}
\title{Generates the partitions by overlaying rasters of the coarse and fine polygons}
\usage{
rasterPartition(polyCoarse, polyFine, cellsCoarse = 40, cellsFine = 400, bw,
  ncores = 2, idFile = paste(tempfile(), "Id.grd", sep = ""),
  offsetFile = paste(tempfile(), "Offset.grd", sep = ""), verbose = FALSE)
}
\arguments{
\item{\code{polyCoarse}}{Spatial polygons of case data}

\item{\code{polyFine}}{Spatial polygons of population data}

\item{\code{bw}}{Vector of bandwidths}

\item{\code{cellsCoarse}}{Horizontal/vertical resolution of raster applied to coarse polygons}

\item{\code{cellsFine}}{Horizontal/vertical resolution of raster applied to fine polygons}

\item{\code{ncores}}{Number of cores/threads for parallel processing}

\item{\code{idFile}}{Filename (must have .grd extension) of the raster of partitions}

\item{\code{offsetFile}}{Filename (must have .grd extension) of the rasters of smoothed offsets}
}
\value{
The \code{rasterPartition} function returns a list containing the raster of the coarse polygons, 
 raster stacks of the partitions and offsets, 
 focal weight matrix, and 
 the coarse spatial polygons.
}
\description{
The \code{rasterPartition} function first rasterizes the coarse and fine spatial polygons based on their respective input resolutions, and then, 
 overlays these rasters to generate the raster of partitions of the local-EM algorithm. 
 It also applies the kernel smoothing function with input bandwidths to the expected counts of the fine polygons to obtain the smoothed offsets (i.e., smoothed expected counts / cell area) of the partitions.
}
\details{
After using the \code{rasterPartition} function, the fine raster is a raster stack containing the IDs for the partitions created by overlaying the coarse and fine rasters. 
 The offset raster is a raster stack containing the offsets of the partitions smoothed with the specified bandwidths. These values represent the denominator of the kernel smoothing matrix.
}
\examples{
data(kentuckyCounty)

data(kentuckyTract)

\dontrun{
lemRaster = rasterPartition(polyCoarse = kentuckyCounty, polyFine = kentuckyTract, 
                   cellsCoarse = 40, cellsFine = 400, 
                   bw = c(10, 15, 20, 25) * 1000, 
                   ncores = 4, 
                   idFile = 'id.grd', offsetFile = 'offset.grd')
}

}

