% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterPartFun.R
\name{rasterPartition}
\alias{rasterPartition}
\title{Generates the partitions by overlaying rasters of spatial polygons}
\usage{
rasterPartition(polyCoarse, polyFine, cellsCoarse, cellsFine, bw,
  focalSize = NULL, xv = NULL, ncores = 1, path = getwd(), idFile,
  offsetFile, verbose = FALSE)
}
\arguments{
\item{polyCoarse}{Spatial polygons of case data}

\item{polyFine}{Spatial polygons of population data}

\item{cellsCoarse}{Minimum resolution for rasterization of case data for numerical accuracy of smoothing matrix}

\item{cellsFine}{Minimum resolution for rasterization of population data for numerical integration of smoothing matrix}

\item{bw}{Vector of bandwidths}

\item{focalSize}{Distance to truncate Gaussian kernel, default is 2.5 times largest bandwidth}

\item{xv}{(Optional) Number of cross-validation sets, or matrix where rows are coarse polygons and columns are cross-validation sets}

\item{ncores}{Number of cores/threads for parallel processing}

\item{path}{Folder to store raster data}

\item{idFile}{Filename (must have .grd extension) of the raster of partitions}

\item{offsetFile}{Filename (must have .grd extension) of the rasters of smoothed offsets}

\item{verbose}{Verbose output}
}
\value{
The \code{rasterPartition} function returns a list containing the raster of the case data, raster stacks of the partitions and offsets, focal weight matrix of the Gaussian kernel, and the input coarse polygons.
}
\description{
The \code{rasterPartition} function first rasterizes the spatial polygons of the case and population data based on their respective input resolutions, and then, overlays these rasters to generate the raster of partitions of the local-EM algorithm. It also applies the kernel smoothing function with input bandwidths to the expected counts of the fine polygons to obtain the smoothed offsets (i.e., smoothed expected counts per cell area) of the partitions. If cross-validation is specified, smoothed offsets are computed for each cross-validation set.
}
\details{
After using the \code{rasterPartition} function, the partition raster is a raster stack containing the IDs for the partitions created by overlaying the rasterizations of the spatial polygons of case and population data. The offset raster is a raster stack containing the offsets of the partitions smoothed with the specified bandwidths. These values represent the denominator of the kernel smoothing matrix.
}
\examples{
\dontrun{ 
# case and population data
data('kentuckyCounty')
data('kentuckyTract')

# parameters
ncores = 2
cellsCoarse = 8
cellsFine = 100
bw = c(10, 15, 17.5, 20) * 1000
path = 'example'

# rasters of case and population data
lemRaster = rasterPartition(polyCoarse = kentuckyCounty, 
							polyFine = kentuckyTract, 
							cellsCoarse = cellsCoarse, 
							cellsFine = cellsFine, 
							bw = bw, 
							ncores = ncores, 
							path = path, 
							idFile = 'lemId.grd', 
							offsetFile = 'lemOffsets.grd', 
							verbose = TRUE)
}

}
