% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemEstFun.R
\name{riskEst}
\alias{riskEst}
\title{Computes the relative risk estimation on the raster of fine polygons}
\usage{
riskEst(x, lemObjects, bw, tol = 1e-06, maxIter = 2000, ncores = 1,
  filename = "")
}
\arguments{
\item{x}{Spatial polygons of case data}

\item{lemObjects}{List of arrays for the smoothing matrix and
raster stacks for the partition and smoothed offsets}

\item{bw}{Numeric value of bandwidth}

\item{tol}{Tolerance for convergence}

\item{maxIter}{Maximum number of iterations}

\item{ncores}{Number of cores/threads for parallel processing}

\item{filename}{Passed to writeRaster}

\item{verbose}{Verbose output}
}
\value{
The \code{riskEst} function returns a the raster of risk estimations for input bandwidth.
}
\description{
The \code{riskEst} function computes the estimations of the relative risk on the cells of the fine raster.
}
\details{
After using the \code{riskEst} function, the raster of risk estimations is done on cells of the raster on the fine polygons.
}
\examples{
data(kentuckyCounty)

data(kentuckyTract)

\dontrun{
lemRaster = rasterPartition(polyCoarse = kentuckyCounty, polyFine = kentuckyTract,
                   cellsCoarse = 6, cellsFine = 100,
                   bw = c(10, 15, 20, 25) * 1000,
                   ncores = 4,
                   idFile = 'id.grd', offsetFile = 'offset.grd')

lemSmoothMat = smoothingMatrix(rasterObjects = lemRaster,
                   ncores = 4)

lemCv = lemXv(x = kentuckyCounty,
                   lemObjects = lemSmoothMat,
                   ncores = 4)
bestBw = lemCv$bw[which.min(lemCv$cv)]

lemRisk = riskEst(x = kentuckyCounty,
                   lemObjects = lemSmoothMat,
                   bw = bestBw)

plot(lemRisk)
}

}

