% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothMatFun.R
\name{smoothingMatrix}
\alias{smoothingMatrix}
\title{Generates the smoothing matrix for the partitions and bandwidths}
\usage{
smoothingMatrix(rasterObjects, ncores = 2, verbose = FALSE)
}
\arguments{
\item{\code{rasterObjects}}{Raster objects of partitions and smoothed offsets}

\item{\code{ncores}}{Number of cores}
}
\value{
The \code{smoothingMatrix} function returns a list containing an array of smoothing matrix entries, and 
 the input rasters of partitions and smoothed offsets.
}
\description{
The \code{smoothingMatrix} function computes the entries of the smoothing matrix for the local-EM algorithm for the input bandwidths based on 
 the partitions created by rasterizing the coarse and fine polygons, and 
 their smoothed offsets created by applying the kernel smoothing function with input bandwidths.
}
\details{
After using the \code{smoothingMatrix} function, the smoothing matrix is an array containing the integrated kernel smoothing entries of the partitions divided by the integrated kernel smoothing entries of the study region for each specified bandwidth.
}
\examples{
data(kentuckyCounty)

data(kentuckyTract)

\dontrun{
lemRaster = rasterPartition(polyCoarse = kentuckyCounty, polyFine = kentuckyTract, 
                   cellsCoarse = 40, cellsFine = 400, 
                   bw = c(10, 15, 20, 25) * 1000, 
                   ncores = 4, 
                   idFile = 'id.grd', offsetFile = 'offset.grd')
                   
lemSmoothMat = smoothingMatrix(rasterObjects = lemRaster, 
                   ncores = 4) 
}

}

