% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothMatFun.R
\name{smoothingMatrix}
\alias{smoothingMatrix}
\title{Generates the smoothing matrix for the partitions and bandwidths}
\usage{
smoothingMatrix(rasterObjects, ncores = 1, path = getwd(), filename,
  verbose = FALSE)
}
\arguments{
\item{rasterObjects}{Raster stacks for partitions and smoothed offsets}

\item{ncores}{Number of cores/threads for parallel processing}

\item{path}{Folder to store raster data}

\item{filename}{Filename (must have .grd extension) of the entries of smoothing matrix}

\item{verbose}{Verbose output}
}
\value{
The \code{smoothingMatrix} function returns a list containing an array of smoothing matrix entries, and the input rasters of partitions and smoothed offsets.
}
\description{
The \code{smoothingMatrix} function computes the entries of the smoothing matrix for the local-EM algorithm for the input bandwidths based on the partitions created by rasterizing the spatial polygons of case and population data, and their smoothed offsets created by applying the kernel smoothing function with input bandwidths.
}
\details{
After using the \code{smoothingMatrix} function, the smoothing matrix is an array containing the integrated kernel smoothing entries of the partitions divided by the integrated kernel smoothing entries of the study region for each specified bandwidth.
}
\examples{
\dontrun{ 
# case and population data
data('kentuckyCounty')
data('kentuckyTract')

# parameters
ncores = 2
cellsCoarse = 8
cellsFine = 100
bw = c(10, 15, 17.5, 20) * 1000
path = 'example'

# rasters of case and population data
lemRaster = rasterPartition(polyCoarse = kentuckyCounty, 
							polyFine = kentuckyTract, 
							cellsCoarse = cellsCoarse, 
							cellsFine = cellsFine, 
							bw = bw, 
							ncores = ncores, 
							path = path, 
							idFile = 'lemId.grd', 
							offsetFile = 'lemOffsets.grd', 
							verbose = TRUE)

# smoothing matrix
lemSmoothMat = smoothingMatrix(rasterObjects = lemRaster, 
								ncores = ncores, 
								path = path, 
								filename = 'lemSmoothMat.grd', 
								verbose = TRUE)
}

}
